+++ src/main/java/com/lawProject/SSL/global/security/filter/JwtAuthenticationProcessingFilter.java
@@ -2,7 +2,7 @@ package com.lawProject.SSL.global.security.filter;
 
 import com.lawProject.SSL.domain.token.repository.BlacklistedTokenRepository;
 import com.lawProject.SSL.domain.token.repository.RefreshTokenRepository;
-import com.lawProject.SSL.domain.user.dao.UserRepository;
+import com.lawProject.SSL.domain.user.repository.UserRepository;
 import com.lawProject.SSL.domain.user.model.User;
 import com.lawProject.SSL.global.util.JwtUtil;
 import jakarta.servlet.FilterChain;
